Instant_Neural_Graphics_Primitives_with_a_Multiresolution_Hash_Encoding/requirements.txt
numpy==1.26.0
pillow==10.0.1
torch==1.13.0
tqdm==4.66.1